requirements.txt
mcp
docker
minio
python-dotenv
requests
pandas
numpy
openpyxl
python-docx

